requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
black==24.4.2
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
distlib==0.3.8
emoji==2.12.1
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
importlib_metadata==7.2.1
libcst==1.4.0
lunchable==1.3.3
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.7.1
pycln==2.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tomlkit==0.12.5
typer==0.12.3
types-emoji==2.1.0.3
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
virtualenv==20.26.3
zipp==3.19.2
